pyproject.toml
[tool.poetry]
name = "docsplit"
version = "0.1.0"
description = "Automated document splitting and archival from batch scans"
authors = ["Brian Gehrke"]
readme = "README.md"
packages = [{include = "docsplit", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pymupdf = "^1.24.0"
pdf2image = "^1.17.0"
pytesseract = "^0.3.10"
ollama = "^0.4.5"
pydantic = "^2.10.0"
pyyaml = "^6.0"
pillow = "^11.0.0"
rapidfuzz = "^3.10.0"
opencv-python = "^4.13.0.92"
numpy = "^2.4.3"
tenacity = "^9.1.4"
flask = "^3.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^6.0.0"
black = "^24.0.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[tool.poetry.scripts]
docsplit = "docsplit.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/docsplit --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
